+++ Assets/Scripts/Card/EffectSystem.cs
@@ -1,85 +1,89 @@
 using UnityEngine;
 using System.Collections.Generic;
 
-public class EffectSystem
+namespace CreatorKousien.Effect
 {
-    struct RegisterEffectData
+    public class EffectSystem
     {
-        public EffectData data;
-        public int currentDuration;
-
-        public RegisterEffectData(EffectData effectData)
+        struct RegisterEffectData
         {
-            this.data = effectData;
-            currentDuration = effectData.Duration;
-        }
-    }
-    //! エフェクト情報
-    EffectDataBase _effectDataBase;
-    //! 
-    List<RegisterEffectData> _registerEffects;
+            public EffectData data;
+            public int currentDuration;
 
-    public EffectSystem(EffectDataBase effectDataBase)
-    {
-        _registerEffects = new List<RegisterEffectData>();
-        _effectDataBase = effectDataBase;
-    }
-
-    public void RegisterData(int effectID)
-    {
-        // 効果を取得
-        EffectData registerData = _effectDataBase.GetEffect(effectID);
-        // 登録
-        _registerEffects.Add(new RegisterEffectData(registerData));
-    }
-
-    public void ApplyAllEffect()
-    {
-        int i = 0;
-        List<int> deleteEffectCounts = new List<int>();
-        foreach (var effect in _registerEffects)
-        {
-            // 遅延中だった場合は処理しない
-            if (effect.currentDuration > 0)
+            public RegisterEffectData(EffectData effectData)
             {
-                Debug.Log("効果ID" + effect.data.EffectID + "は残り待機ターン" +　effect.currentDuration + "ターンなのでスキップします");
+                this.data = effectData;
+                currentDuration = effectData.Duration;
+            }
+        }
+        //! エフェクト情報
+        EffectDataBase _effectDataBase;
+        //! 
+        List<RegisterEffectData> _registerEffects;
+
+        public EffectSystem(EffectDataBase effectDataBase)
+        {
+            _registerEffects = new List<RegisterEffectData>();
+            _effectDataBase = effectDataBase;
+        }
+
+        public void RegisterData(int effectID)
+        {
+            // 効果を取得
+            EffectData registerData = _effectDataBase.GetEffect(effectID);
+            // 登録
+            _registerEffects.Add(new RegisterEffectData(registerData));
+        }
+
+        public void ApplyAllEffect()
+        {
+            int i = 0;
+            List<int> deleteEffectCounts = new List<int>();
+            foreach (var effect in _registerEffects)
+            {
+                // 遅延中だった場合は処理しない
+                if (effect.currentDuration > 0)
+                {
+                    Debug.Log("効果ID" + effect.data.EffectID + "は残り待機ターン" + effect.currentDuration + "ターンなのでスキップします");
+                    i++;
+                    continue;
+                }
+                // 処理を反映
+                if (ApplyEffect(effect.data))
+                {
+                    deleteEffectCounts.Add(i);
+                }
+                else
+                {
+                    Debug.Log("[EffectSystem]" + effect.data.EffectName + "の反映に失敗しました");
+                }
+
                 i++;
-                continue;
+
             }
-            // 処理を反映
-            if(ApplyEffect(effect.data))
+            // 使用した効果を削除
+            for (int j = deleteEffectCounts.Count - 1; j >= 0; j--)
             {
-                deleteEffectCounts.Add(i);
+                _registerEffects.RemoveAt(deleteEffectCounts[j]);
             }
-            else
+
+        }
+
+        public void DurationUpdate(int turn)
+        {
+            for (int i = 0; i < _registerEffects.Count; i++)
             {
-                Debug.Log("[EffectSystem]" + effect.data.EffectName + "の反映に失敗しました");
+                RegisterEffectData effect = _registerEffects[i];
+                effect.currentDuration -= turn;
+                _registerEffects[i] = effect;
             }
-
-            i++;
-
         }
-        // 使用した効果を削除
-        for(int j = deleteEffectCounts.Count - 1; j >= 0; j--)
+
+        bool ApplyEffect(EffectData data)
         {
-            _registerEffects.RemoveAt(deleteEffectCounts[j]);
+            Debug.Log("ID" + data.EffectID + "効果名" + data.EffectName + "実行！");
+            return true;
         }
-
-    }
-
-    public void DurationUpdate(int turn)
-    {
-        for(int i = 0; i < _registerEffects.Count; i++)
-        {
-            RegisterEffectData effect = _registerEffects[i];
-            effect.currentDuration -= turn;
-            _registerEffects[i] = effect;
-        }
-    }
-
-    bool ApplyEffect(EffectData data)
-    {
-        Debug.Log("ID" + data.EffectID + "効果名" + data.EffectName + "実行！");
-        return true;
     }
 }
+
