+++ Assets/Scripts/Command/Card.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0f9cc8fb7bf89ea44ba72860f0c408f3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
